requirements.txt
crewai
crewai_tools
yfinance
python-dotenv
